requirements.txt
streamlit 
python-dotenv
google-generativeai
pyPDF2
chromadb
langchain
